+++ src/components/employees/Employees.js
@@ -2,6 +2,7 @@ import React from 'react'
 import axios from '../../config/axios'
 import { Link } from 'react-router-dom'
 import Table from './Tables'
+import Swal from 'sweetalert2'
 
 class EmployeeList extends React.Component {
     constructor () {
@@ -26,18 +27,41 @@ class EmployeeList extends React.Component {
     }
 
     handleRemove = (data) => {
-        axios.delete(`/employees/${data}`)
-        .then(response => {
-            if (response.data._id) {
-                this.setState(prevState => {
-                    return {
-                        employees : prevState.employees.filter(empl => empl._id != response.data._id)
-                    }
-                })
-            } else {
-                alert (response.data.message)
+        Swal.fire({
+            title: 'Are you sure?',
+            text: "You won't be able to revert this!",
+            icon: 'warning',
+            showCancelButton: true,
+            confirmButtonColor: '#3085d6',
+            cancelButtonColor: '#d33',
+            confirmButtonText: 'Yes, delete it!'
+          }).then((result) => {
+            if (result.value) {
+                axios.delete(`/employees/${data}`)
+                    .then(response => {
+                        if (response.data._id) {
+                            this.setState(prevState => {
+                                return {
+                                    employees : prevState.employees.filter(empl => empl._id != response.data._id)
+                                }
+                            })
+                            Swal.fire(
+                                'Deleted!',
+                                'Your file has been deleted.',
+                                'success'
+                              )
+                        } else {
+                            Swal.fire(
+                                'Oops!',
+                                'Something went wrong',
+                                'error'
+                              )
+                        }
+                    }) 
+             
             }
-        }) 
+          })
+        
     }
 
     render () {
